refactor: serve icons from their own design system barrel
