+++ FateConnect/Web/src/design-system/components/ThemeToggleButton/index.tsx
@@ -1,7 +1,7 @@
 import IconButton from '@mui/material/IconButton';
 
 import { useThemeMode } from '../../ThemeProvider/context/ThemeModeContext';
-import { DarkModeIcon, LightModeIcon } from '../../ui';
+import { DarkModeIcon, LightModeIcon } from '../../icons';
 
 const LABEL_PARA_ESCURO = 'Ativar tema escuro';
 const LABEL_PARA_CLARO = 'Ativar tema claro';
