+++ FateConnect/Web/src/design-system/icons.ts
@@ -1,0 +1,30 @@
+/**
+ * Ícones expostos pelo design system, separados dos componentes.
+ *
+ * A aplicação importa daqui — `import { AddIcon } from '@design-system/icons'` —,
+ * nunca de `@mui/icons-material`: trocar a biblioteca de ícones fica sendo uma
+ * mudança neste arquivo, não uma varredura pelo app.
+ */
+export { default as AccessTimeIcon } from '@mui/icons-material/AccessTime';
+export { default as AddIcon } from '@mui/icons-material/Add';
+export { default as ArrowBackIcon } from '@mui/icons-material/ArrowBack';
+export { default as CalendarTodayIcon } from '@mui/icons-material/CalendarToday';
+export { default as DarkModeIcon } from '@mui/icons-material/DarkMode';
+export { default as DeleteIcon } from '@mui/icons-material/Delete';
+export { default as DirectionsCarIcon } from '@mui/icons-material/DirectionsCar';
+export { default as EditIcon } from '@mui/icons-material/Edit';
+export { default as EmailIcon } from '@mui/icons-material/Email';
+export { default as FilterAltIcon } from '@mui/icons-material/FilterAlt';
+export { default as GroupsIcon } from '@mui/icons-material/Groups';
+export { default as InfoIcon } from '@mui/icons-material/Info';
+export { default as LightModeIcon } from '@mui/icons-material/LightMode';
+export { default as LocationOnIcon } from '@mui/icons-material/LocationOn';
+export { default as MenuIcon } from '@mui/icons-material/Menu';
+export { default as PhoneIcon } from '@mui/icons-material/Phone';
+export { default as ScheduleIcon } from '@mui/icons-material/Schedule';
+export { default as SearchIcon } from '@mui/icons-material/Search';
+export { default as SecurityIcon } from '@mui/icons-material/Security';
+export { default as VisibilityIcon } from '@mui/icons-material/Visibility';
+export { default as VisibilityOffIcon } from '@mui/icons-material/VisibilityOff';
+
+export type { SvgIconComponent } from '@mui/icons-material';
