+++ FateConnect/Web/src/design-system/ui.ts
@@ -33,29 +33,6 @@ export { default as Typography } from '@mui/material/Typography';
 export type { BoxProps } from '@mui/material/Box';
 export type { ButtonProps } from '@mui/material/Button';
 export type { TypographyProps } from '@mui/material/Typography';
-export type { SvgIconComponent } from '@mui/icons-material';
-
-export { default as AccessTimeIcon } from '@mui/icons-material/AccessTime';
-export { default as AddIcon } from '@mui/icons-material/Add';
-export { default as ArrowBackIcon } from '@mui/icons-material/ArrowBack';
-export { default as CalendarTodayIcon } from '@mui/icons-material/CalendarToday';
-export { default as DarkModeIcon } from '@mui/icons-material/DarkMode';
-export { default as DeleteIcon } from '@mui/icons-material/Delete';
-export { default as DirectionsCarIcon } from '@mui/icons-material/DirectionsCar';
-export { default as EditIcon } from '@mui/icons-material/Edit';
-export { default as EmailIcon } from '@mui/icons-material/Email';
-export { default as FilterAltIcon } from '@mui/icons-material/FilterAlt';
-export { default as GroupsIcon } from '@mui/icons-material/Groups';
-export { default as InfoIcon } from '@mui/icons-material/Info';
-export { default as LightModeIcon } from '@mui/icons-material/LightMode';
-export { default as LocationOnIcon } from '@mui/icons-material/LocationOn';
-export { default as MenuIcon } from '@mui/icons-material/Menu';
-export { default as PhoneIcon } from '@mui/icons-material/Phone';
-export { default as ScheduleIcon } from '@mui/icons-material/Schedule';
-export { default as SearchIcon } from '@mui/icons-material/Search';
-export { default as SecurityIcon } from '@mui/icons-material/Security';
-export { default as VisibilityIcon } from '@mui/icons-material/Visibility';
-export { default as VisibilityOffIcon } from '@mui/icons-material/VisibilityOff';
 
 /** Componentes de data. O locale vem do `DateLocalizationProvider`. */
 export { DateCalendar } from '@mui/x-date-pickers/DateCalendar';
