+++ FateConnect/Web/src/pages/Home/components/LandingServices/constants/index.ts
@@ -4,7 +4,7 @@ import {
   SearchIcon,
   SecurityIcon,
   type SvgIconComponent,
-} from '@design-system';
+} from '@design-system/icons';
 
 export type ServiceCard = { title: string; description: string; Icon: SvgIconComponent };
